+++ app/src/commonMain/kotlin/it/vercruysse/lemmyapi/v0x18/datatypes/ListCommunities.kt
@@ -9,7 +9,7 @@ internal data class ListCommunities(
     val type_: ListingType? /* "All" | "Local" | "Subscribed" */ = null,
     val sort: SortType? /* "Active" | "Hot" | "New" | "Old" | "TopDay" | "TopWeek" | "TopMonth" | "TopYear" | "TopAll" | "MostComments" | "NewComments" | "TopHour" | "TopSixHour" | "TopTwelveHour" | "TopThreeMonths" | "TopSixMonths" | "TopNineMonths" */ = null,
     val show_nsfw: Boolean? = null,
-    val page: Int? = null,
-    val limit: Int? = null,
+    val page: Long? = null,
+    val limit: Long? = null,
     val auth: String? = null,
 )
