+++ app/src/commonMain/kotlin/it/vercruysse/lemmyapi/v0x19/datatypes/GetPersonDetails.kt
@@ -8,8 +8,8 @@ data class GetPersonDetails(
     val person_id: PersonId? = null,
     val username: String? = null,
     val sort: SortType? /* "Active" | "Hot" | "New" | "Old" | "TopDay" | "TopWeek" | "TopMonth" | "TopYear" | "TopAll" | "MostComments" | "NewComments" | "TopHour" | "TopSixHour" | "TopTwelveHour" | "TopThreeMonths" | "TopSixMonths" | "TopNineMonths" | "Controversial" | "Scaled" */ = null,
-    val page: Int? = null,
-    val limit: Int? = null,
+    val page: Long? = null,
+    val limit: Long? = null,
     val community_id: CommunityId? = null,
     val saved_only: Boolean? = null,
 )
