+++ app/src/commonMain/kotlin/it/vercruysse/lemmyapi/v0x19/datatypes/ListRegistrationApplications.kt
@@ -5,6 +5,6 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class ListRegistrationApplications(
     val unread_only: Boolean? = null,
-    val page: Int? = null,
-    val limit: Int? = null,
+    val page: Long? = null,
+    val limit: Long? = null,
 )
