+++ app/src/commonMain/kotlin/it/vercruysse/lemmyapi/v0x19/datatypes/PersonAggregates.kt
@@ -5,6 +5,6 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class PersonAggregates(
     val person_id: PersonId,
-    val post_count: Int,
-    val comment_count: Int,
+    val post_count: Long,
+    val comment_count: Long,
 )
